# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ServiceCreate": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "maxLength": 3,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deposit_minor_units": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Minor Units"
          },
          "deposit_required": {
            "default": false,
            "title": "Deposit Required",
            "type": "boolean"
          },
          "duration_minutes": {
            "maximum": 480.0,
            "minimum": 1.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "price_minor_units": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Minor Units"
          }
        },
        "required": [
          "name",
          "duration_minutes"
        ],
        "title": "ServiceCreate",
        "type": "object"
      },
      "ServiceRead": {
        "properties": {
          "business_id": {
            "title": "Business Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deposit_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Minor Units"
          },
          "deposit_required": {
            "title": "Deposit Required",
            "type": "boolean"
          },
          "duration_minutes": {
            "title": "Duration Minutes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price_minor_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Minor Units"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "business_id",
          "name",
          "duration_minutes",
          "is_active",
          "price_minor_units",
          "currency",
          "deposit_required",
          "deposit_minor_units",
          "created_at"
        ],
        "title": "ServiceRead",
        "type": "object"
      },
      "ServiceUpdate": {
        "properties": {
          "currency": {
            "anyOf": [
              {
                "maxLength": 3,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deposit_minor_units": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Minor Units"
          },
          "deposit_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Required"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "maximum": 480.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "price_minor_units": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Minor Units"
          }
        },
        "title": "ServiceUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "description": "Use the access token returned by login or refresh. Send it as `Authorization: Bearer <access_token>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Appointment Voice SaaS API \u2014 multi-tenant booking and availability backend built on a production FastAPI foundation (JWT auth, admin workflows, audit logs, Redis rate limiting, object storage). Salon domain: businesses, staff, services, hours, availability slots, bookings with DB-level double-booking protection. IVR/SMS/calendar: planned.",
    "summary": "Multi-tenant booking and availability backend for local service businesses",
    "title": "Appointment Voice SaaS API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/businesses/{business_id}/services": {
      "get": {
        "operationId": "list_services_endpoint_api_v1_businesses__business_id__services_get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceRead"
                  },
                  "title": "Response List Services Endpoint Api V1 Businesses  Business Id  Services Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Services Endpoint",
        "tags": [
          "services"
        ]
      },
      "post": {
        "operationId": "create_service_endpoint_api_v1_businesses__business_id__services_post",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Service Endpoint",
        "tags": [
          "services"
        ]
      }
    },
    "/api/v1/businesses/{business_id}/services/{service_id}": {
      "delete": {
        "operationId": "delete_service_endpoint_api_v1_businesses__business_id__services__service_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Service Endpoint",
        "tags": [
          "services"
        ]
      },
      "get": {
        "operationId": "get_service_endpoint_api_v1_businesses__business_id__services__service_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Service Endpoint",
        "tags": [
          "services"
        ]
      },
      "patch": {
        "operationId": "update_service_endpoint_api_v1_businesses__business_id__services__service_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "business_id",
            "required": true,
            "schema": {
              "title": "Business Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Service Endpoint",
        "tags": [
          "services"
        ]
      }
    },
...
  },
...
}